nptime.nptime.__add__: [Fix] Play nicely with subclasses
If `nptime` is subclassed, the current implementation of `__add__` will return
an instance of `nptime.nptime` instead of an instance of the subclass. This is
because the `from_time` call is made on `nptime` instead of on `self`. This
commit changes this so that current functionality is maintained, but adding a
subclass of `nptime.nptime` and a `timedelta` returns an instance of the
`subclass`.